:art: refactor : 여러 DAO에서 사용할 수 있는 connection을 따로 분리
